api design updated
