+++ backend/src/dating/chats/exceptions.py
@@ -1,10 +1,6 @@
 from ..users.exceptions import UserHasNoPermission
 
 
-class UserNotInChat(ValueError):
-    pass
-
-
 class ChatNotFound(ValueError):
     pass
